+++ dev/changelog.md
@@ -1,3 +1,4 @@
 # pyABF 2.3.0
 * Improved support for user lists (#110) _Thanks @haganenoneko_
-* ABFs with invalid dates no longer display a warning when they are loaded
+* ABFs with invalid dates no longer display a warning when they are loaded (#113)
+* ABF and ATF now can also be instantiated with a `pathlib.Path` instead of just `str` (#112)